[Win32] Actually fix win32 corruption and fix a memory leak.

+++ main/src/addins/WindowsPlatform/structs.cs
@@ -169,7 +169,7 @@ namespace CustomControls.OS
     {
         public IntPtr hIcon;
         public int iIcon;
-        public ulong dwAttributes;
+        public uint dwAttributes;
         [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 260)]
         public string szDisplayName;
         [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 80)]
